Error when deleting vhs entry

+++ vhs-backend/src/vhs/vhs.service.ts
@@ -59,9 +59,11 @@ export class VhsService {
     }
 
     await this.vhsRepository.delete(id);
-    fs.unlink(vhs.thumbnail, (error) => {
-      if (error) {
-        throw error;
+    if(vhs.thumbnail){
+      fs.unlink(vhs.thumbnail, (error) => {
+        if (error) {
+          throw error;
+        }
       }
     });
   }
